갤러리, 페이지네이션 완성
갤러리 저장 기능, 페이지네이션 기능, 공유 취소 완성

+++ src/main/java/com/steepcliff/thinkboom/brainWriting/dto/bwRoom/BwRoomResponseDto.java
@@ -4,8 +4,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.UUID;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -15,9 +13,9 @@ public class BwRoomResponseDto {
 
     private Integer headCount;
 
-    private Long time;
+    private Integer time;
 
-    public BwRoomResponseDto(String roomId,Integer headCount, Long time) {
+    public BwRoomResponseDto(String roomId,Integer headCount, Integer time) {
         this.roomId = roomId;
         this.headCount = headCount;
         this.time = time;
